server/requirements.txt
flask==2.3.3
openai==1.3.0
requests==2.31.0
Pillow==10.0.0
gradio_client==0.6.1
python-dotenv==1.0.0
boto3==1.34.63